requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.10.0
apscheduler==3.11.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
distro==1.9.0
docstring-parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy-extensions==1.1.0
natsort==8.4.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.103.0
opencv-python==4.12.0.88
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
proglog==0.1.12
prompt-toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlalchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.11
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.50.0
triton==3.4.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.180
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
